Move fetch_api_description to m.client.bones.helpers.
